Include running the harness/setup.py test
Skip-unit-tests: true
Test-tag: pr test_agent_failure_isolation ServerRankFailure control pool test_cleanup_procs HarnessSetupTest

Required-githooks: true
