+++ src/tests/ftest/harness/setup.py
@@ -12,6 +12,14 @@ class HarnessSetupTest(TestWithServers):
     :avocado: recursive
     """
 
+    def run_test(self):
+        """Run the test."""
+        if self.server_managers[0].storage_prepare_timeout.value != 60:
+            self.fail("Storage prepare was not set correctly from the test yaml")
+        if self.server_managers[0].storage_format_timeout.value != 60:
+            self.fail("Storage format was not set correctly from the test yaml")
+        self.log.info("Test passed!")
+
     def test_setup(self):
         """Verify the TestWithServers.setUp() method.
 
@@ -22,8 +30,16 @@ class HarnessSetupTest(TestWithServers):
         :avocado: tags=harness
         :avocado: tags=HarnessSetupTest,test_setup
         """
-        self.assertEqual(self.server_managers[0].storage_prepare_timeout.value, 60,
-                         "FAILED: storage prepare was not set correctly from the yaml")
-        self.assertEqual(self.server_managers[0].storage_format_timeout.value, 60,
-                         "FAILED: storage format was not set correctly from the yaml")
-        self.log.info("Test passed!")
+        self.run_test()
+
+    def test_setup_hw(self):
+        """Verify the TestWithServers.setUp() method.
+
+        Useful for setting up the /etc/daos/daos_server.yml files on multiple hosts.
+
+        :avocado: tags=all
+        :avocado: tags=hw,medium,large
+        :avocado: tags=harness
+        :avocado: tags=HarnessSetupTest,test_setup_hw
+        """
+        self.run_test()
