--- src/tests/ftest/harness/setup_hw.yaml
@@ -1,13 +1,0 @@
-hosts:
-  test_servers: 1
-  test_clients: 1
-timeout: 120
-server_config:
-  name: daos_server
-  engines_per_host: 1
-  engines:
-    0:
-      storage: auto
-server_manager:
-  storage_prepare_timeout: 60
-  storage_format_timeout: 60
